requirements.txt
aiohttp
asyncpg
loguru
PyYAML
cerberus
psycopg2-binary
yoyo-migrations